requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
jupyter
beautifulsoup4
requests-html